Move the build over to a dual CommonJS and ESM build
