+++ src/instrumentation/fetch.ts
@@ -10,11 +10,11 @@ import {
 	SpanStatusCode,
 } from '@opentelemetry/api'
 import { SemanticAttributes } from '@opentelemetry/semantic-conventions'
-import { Initialiser, getActiveConfig, setConfig } from '../config'
-import { wrap } from './wrap'
-import { instrumentEnv } from './env'
-import { exportSpans, proxyExecutionContext } from './common'
-import { ResolvedTraceConfig } from '../types'
+import { Initialiser, getActiveConfig, setConfig } from '../config.js'
+import { wrap } from '../wrap.js'
+import { instrumentEnv } from './env.js'
+import { exportSpans, proxyExecutionContext } from './common.js'
+import { ResolvedTraceConfig } from '../types.js'
 
 export type IncludeTraceContextFn = (request: Request) => boolean
 export interface FetcherConfig {
