+++ src/tracer.ts
@@ -12,8 +12,8 @@ import { sanitizeAttributes } from '@opentelemetry/core'
 import { Resource } from '@opentelemetry/resources'
 import { SpanProcessor, RandomIdGenerator, ReadableSpan, SamplingDecision } from '@opentelemetry/sdk-trace-base'
 
-import { SpanImpl } from './span'
-import { getActiveConfig } from './config'
+import { SpanImpl } from './span.js'
+import { getActiveConfig } from './config.js'
 
 export class WorkerTracer implements Tracer {
 	private readonly _spanProcessor: SpanProcessor
